Fix wrong-result and crash bugs found in 2026-08-19 deep review (B1-B13, C1-C13)
Addresses the master-branch triage in docs/20260819_fablereview.md: decoy
NL-ladder early exit, FI/PI stale start-pos and precursor-NL staleness,
mod-count-cap bypasses (FASTA and FI/PI), PEFF mass-compare/duplicate-check
bugs, .idx header restore now also derives bNtermMod/bCtermMod/
bVarTermModSearch (B5), AScorePro/param-name mismatches, enzyme-validation
dead code, FI/PI dedup mod-filter loss, several scoring-consistency
mismatches, and the crashes/memory-safety cluster (off-by-one heap
read/writes, OOM use-after-free, PEFF/.idx/MSP/mzIdentML robustness against
corrupt input, and related overflow/UB fixes).

+++ CometSearch/CombinatoricsUtils.h
@@ -23,8 +23,11 @@ public:
    ~CombinatoricsUtils();
 
    static int** makeCombinations(int n, int r, int count);
-   static int nChooseK(int n, int k);
-   static int getCombinationCount(int n, int k);
+   // int64_t, not int: C(44,10) alone is ~2.48 billion, already past INT_MAX, and
+   // initBinomialCoefficients()'s default (n=50, k=10) builds rows well past that -- see the
+   // .cpp for the full explanation.
+   static long long nChooseK(int n, int k);
+   static long long getCombinationCount(int n, int k);
    static void initBinomialCoefficients(const int n, const int k);
 };
